+++ zerver/lib/cache_helpers.py
@@ -9,10 +9,11 @@ from typing import Any, Dict, Callable, Tuple
 from django.conf import settings
 from zerver.models import Message, UserProfile, Stream, get_stream_cache_key, \
     Recipient, get_recipient_cache_key, Client, get_client_cache_key, \
-    Huddle, huddle_hash_cache_key, to_dict_cache_key_id
+    Huddle, huddle_hash_cache_key
 from zerver.lib.cache import cache_with_key, cache_set, \
     user_profile_by_email_cache_key, user_profile_by_id_cache_key, \
-    get_remote_cache_time, get_remote_cache_requests, cache_set_many
+    get_remote_cache_time, get_remote_cache_requests, cache_set_many, \
+    to_dict_cache_key_id
 from django.utils.importlib import import_module
 from django.contrib.sessions.models import Session
 import logging
